Extract common fixtures into contest.py
