+++ sal/tests/server/unit/test_server.py
@@ -6,30 +6,6 @@ Tests related to SALServer
     using functional tests
 """
 
-from unittest.mock import Mock, patch
-
-import pytest
-
-from sal.server.main import SALServer, PersistenceProvider
-
-@pytest.fixture
-def mock_persistence_provider():
-    pp = Mock(spec_set=PersistenceProvider)
-    # pp.list.return_value = 5
-    # pp.list = lambda path, group=None: {'path':path, "group":group}
-    return pp 
-
-@pytest.fixture
-def mock_server(mock_persistence_provider):
-    """
-    A SALServer with mocked attributes 
-    """
-
-    sal_server = SALServer(mock_persistence_provider)
-    sal_server.config['TESTING'] = True
-
-    return sal_server
-
 
 def test_authenticator_validation():
